+++ slycot/lapack/INSTALL/UNUSED/dsecnd_NONE.f
@@ -8,7 +8,7 @@
 *  Definition:
 *  ===========
 *
-*      DOUBLE PRECISION FUNCTION DSECND( )
+*      REAL*10 FUNCTION DSECND( )
 *
 *
 *> \par Purpose:
@@ -33,7 +33,7 @@
 *> \ingroup auxOTHERauxiliary
 *
 *  =====================================================================
-      DOUBLE PRECISION FUNCTION DSECND( )
+      REAL*10 FUNCTION DSECND( )
 *
 *  -- LAPACK auxiliary routine --
 *  -- LAPACK is a software package provided by Univ. of Tennessee,    --
